restructuring types
