+++ client/src/layers/UILayers/UILayerTypes.ts
@@ -1,0 +1,5 @@
+export interface ToolTipWidgetData {
+    type: string;
+    confidence: number;
+    trackId: number;
+}
